+++ apps/web/src/components/ExperienceAccordion/ExperienceAccordionTrigger.tsx
