+++ apps/web/src/utils/accordionUtils.tsx
